fastApi/requirements.txt
transformers
python-dotenv
duckduckgo-search==5.1.0
google-serp-api
langchain
langchain-groq
langchain-community
wikipedia
gradio
torch
torchvision
datasets==2.16.0
bitsandbytes
peft
accelerate
Pillow
uvicorn
fastapi
trl